fix DHO not showing up
this was a comma at the end of the JSON before the closing bracket. Some servers were OK with that, others were not. Definitely a bug though, fixed. This is why I like fail early - all servers should actually fail when the JSON is invalid. I used the JSON from cleos so that's why it's longer.