Release/v1.4.3 (#474)
* transfer screen integrates available balance under amount

- bugfixes and error entry handling
- layout for the under the amount elements

* better looks for small screens

* cleanup remove dependency on rate notifier

this becomes a more dumb widget which gets the rate from outside and as text

* build number

* more sensible quantity formatting with 2 trailing zeroes

honestly nobody cares about the last 2 digits - not in a normal case anyway.

* xcode stuff

* remove printout

* add progress bar to avoid jarring transition

* remove unused import

* version

* progress indicator since it takes a long time to generate a QR code

* layout fix, size fix

small devices are only < 320 pixels
using Expanded expresses much more what we want - wrapped text that fills the space.

* fix validation bug

* no validation on changed - the field validates on input but this gets triggered when the currency changes

* consistent number formatting, unit tests for number format

* build 28

* fix incorrect amount - whoops

* line item total fix

* better max guardian message

* fix error in catch

* fix error where endpoint was ignored

* 1.4.2

* different default endpoint

* fix DHO not showing up

this was a comma at the end of the JSON before the closing bracket. Some servers were OK with that, others were not. Definitely a bug though, fixed. This is why I like fail early - all servers should actually fail when the JSON is invalid. I used the JSON from cleos so that's why it's longer.

* part of fthe last commit

something weird happened here with git... this is part of the previius commit

* Revert "Merge branch 'master' into release/v1.4.2"

This reverts commit 4b61ae0779cd651e26255072c7d5e4c168265e4e, reversing
changes made to e62da8abafe1c8eda771a60f0423b4ec00ef64d6.

* Restore QR button

* not full black

* fix version, remove workflow

* Revert "Revert "Merge branch 'master' into release/v1.4.2""

This reverts commit 765494ac14111be6539741932231fe6f77b9234b.